+++ Package.swift
@@ -8,7 +8,6 @@ let package = Package(
         .iOS(.v17),
     ],
     products: [
-        .library(name: "BroadMonetization", targets: ["BroadMonetization"]),
         .library(name: "BroadUIFlows", targets: ["BroadUIFlows"]),
     ],
     dependencies: [
@@ -21,32 +20,20 @@ let package = Package(
             exact: "1.0.0"
         ),
         .package(
-            url: "https://github.com/Swinject/Swinject.git",
-            exact: "2.10.0"
+            url: "https://github.com/BroadApps-official/broad-monetization-ios.git",
+            exact: "1.0.0"
         ),
         .package(
-            url: "https://github.com/adaptyteam/AdaptySDK-iOS.git",
-            exact: "3.17.3"
+            url: "https://github.com/Swinject/Swinject.git",
+            exact: "2.10.0"
         ),
     ],
     targets: [
         .target(
-            name: "BroadMonetization",
-            dependencies: [
-                .product(name: "BroadCore", package: "broad-core-ios"),
-                .product(name: "Swinject", package: "Swinject"),
-                .product(
-                    name: "Adapty",
-                    package: "AdaptySDK-iOS",
-                    condition: .when(platforms: [.iOS])
-                ),
-            ]
-        ),
-        .target(
             name: "BroadUIFlows",
             dependencies: [
                 .product(name: "BroadCore", package: "broad-core-ios"),
-                "BroadMonetization",
+                .product(name: "BroadMonetization", package: "broad-monetization-ios"),
                 .product(name: "Swinject", package: "Swinject"),
             ]
         ),