+++ Sources/BroadMonetization/Infrastructure/Analytics/NoOpMonetizationAnalytics.swift
@@ -1,7 +1,0 @@
-public struct NoOpMonetizationAnalytics: MonetizationAnalyticsProtocol {
-    public init() {}
-
-    public func track(
-        _ event: MonetizationAnalyticsEvent
-    ) async {}
-}