+++ Sources/BroadMonetization/Infrastructure/Entitlements/UnknownEntitlementStatusProvider.swift
@@ -1,7 +1,0 @@
-public struct UnknownEntitlementStatusProvider: EntitlementStatusProviderProtocol {
-    public init() {}
-
-    public func currentStatus() async -> EntitlementStatus {
-        .unknown
-    }
-}